requirements.txt
jupyterlab == 4.2.2
langchain == 0.2.5
openai == 1.34.0
streamlit == 1.35.0
chromadb == 0.5.0
langchain-openai == 0.1.8
langchain-community == 0.2.5
PyPDF2==3.0.1
faiss-cpu
python-dotenv==1.0.0
altair==4
tiktoken==0.4.0
unstructured
langchain-experimental
langgraph